Fix controllers, rename table Order

+++ src/main/java/com/example/demo/dto/response/ClientResponse.java
@@ -9,5 +9,5 @@ public class ClientResponse {
     private Integer id;
     private String fullName;
     private String email;
-    private Integer telNumber;
+    private Long telNumber;
 }
